generalize methods further by creating default method
This commit refactors the methods interface
further with the creation of a DefaultMethod struct
that more specific methods embed to simplify common
tasks between the methods such as getting and
setting the method's target.

This commit also changes the configuration file
to have fetched create the method structs itself
rather than to have viper do that, by changing
the configuration file to accept a list of the
methods with a kind field specifying the type of
the method.

Signed-off-by: Joseph Sawaya <jsawaya@redhat.com>